requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0

# Technical analysis
ta>=0.10.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0

# Jupyter notebooks
jupyter>=1.0.0
nbformat>=5.8.0